BE refactored with new architecture
