--- api/source/resources/user/payload/loginResponse.ts
@@ -1,6 +1,0 @@
-interface LoginResponse {
-  id: string;
-  token: string;
-}
-
-export default LoginResponse;
